requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.0
python-docx==1.1.2
openai==1.79.0
elevenlabs==2.11.0
Pillow==11.2.1
requests==2.32.3